# wp-content/plugins/plugin_estudo/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "WordPress REST API - Plugin Estudo",
    "description": "Documentação completa OpenAPI 3.0 / Swagger de todos os controladores da WordPress REST API (api/v1) do plugin plugin_estudo. Inclui autenticação JWT, configurações, categorias, produtos, posts, páginas e calculadora.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://luizbrogliatto.freedev.app/wp-json",
      "description": "Servidor de Desenvolvimento / Produção"
    }
  ],
...
  "paths": {
...
    "/api/v1/marca": {
      "get": {
        "tags": ["Marcas"],
        "summary": "Listar marcas cadastradas",
        "description": "Retorna a lista paginada de marcas cadastradas na tabela wpo2_marca.",
        "operationId": "listarMarcas",
        "parameters": [
          { "name": "page", "in": "query", "description": "Número da página", "schema": { "type": "integer", "default": 1 }, "example": 1 },
          { "name": "per_page", "in": "query", "description": "Marcas por página", "schema": { "type": "integer", "default": 10 }, "example": 10 },
          { "name": "all", "in": "query", "description": "Retornar todas sem paginação", "schema": { "type": "boolean" }, "example": false }
        ],
        "responses": {
          "200": {
            "description": "Lista de marcas obtida com sucesso",
            "headers": {
              "x-wp-total": { "description": "Total geral de marcas", "schema": { "type": "integer" } },
              "x-wp-totalpages": { "description": "Total de páginas de marcas", "schema": { "type": "integer" } }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Marca" }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Marcas"],
        "summary": "Criar nova marca",
        "description": "Cadastra uma nova marca (campos id e nome).",
        "operationId": "criarMarca",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/MarcaEntrada" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Marca criada com sucesso",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Marca" }
              }
            }
          },
          "400": {
            "description": "Erro na requisição (campo nome obrigatório)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MensagemResposta" }
              }
            }
          }
        }
      }
    },
    "/api/v1/marca/{id}": {
      "get": {
        "tags": ["Marcas"],
        "summary": "Obter marca por ID",
        "description": "Retorna os dados completos de uma marca específica por ID.",
        "operationId": "obterMarcaPorId",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "description": "ID da marca", "schema": { "type": "integer" }, "example": 1 }
        ],
        "responses": {
          "200": {
            "description": "Dados da marca retornados com sucesso",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Marca" }
              }
            }
          },
          "404": {
            "description": "Marca não encontrada",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MensagemResposta" },
                "example": { "message": "Marca não encontrada" }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Marcas"],
        "summary": "Atualizar marca existente",
        "description": "Atualiza os dados de uma marca existente pelo ID.",
        "operationId": "atualizarMarca",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "description": "ID da marca a atualizar", "schema": { "type": "integer" }, "example": 1 }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/MarcaEntrada" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Marca atualizada com sucesso",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Marca" }
              }
            }
          },
          "400": {
            "description": "Falha ao atualizar marca",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MensagemResposta" }
              }
            }
          },
          "404": {
            "description": "Marca não encontrada",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MensagemResposta" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Marcas"],
        "summary": "Excluir marca por ID",
        "description": "Remove a marca do banco de dados pelo ID.",
        "operationId": "excluirMarca",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "description": "ID da marca a excluir", "schema": { "type": "integer" }, "example": 1 }
        ],
        "responses": {
          "200": {
            "description": "Marca excluída com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "Marca excluída com sucesso" },
                    "id": { "type": "integer", "example": 1 }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Falha ao excluir marca",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MensagemResposta" }
              }
            }
          },
          "404": {
            "description": "Marca não encontrada",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MensagemResposta" }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Insira o token JWT gerado pelo endpoint /api/v1/login."
      }
    },
    "schemas": {
      "MensagemResposta": {
        "type": "object",
        "properties": {
          "message": { "type": "string", "example": "Operação executada com sucesso" }
        }
      },
...
      "Marca": {
        "type": "object",
        "description": "Registro da tabela wpo2_marca",
        "properties": {
          "id": {
            "type": "integer",
            "description": "ID único da marca (chave primária)",
            "example": 1
          },
          "nome": {
            "type": "string",
            "description": "Nome da marca",
            "example": "Nike"
          }
        }
      },
      "MarcaEntrada": {
        "type": "object",
        "description": "Payload de entrada para criação ou atualização de marca",
        "required": ["nome"],
        "properties": {
          "nome": {
            "type": "string",
            "description": "Nome da marca",
            "example": "Nike"
          }
        }
      }
    }
  }
}